+++ src/rapporto/github/cli.py
@@ -9,10 +9,12 @@ from rapporto.github.activity import GitHubActivityReport
 from rapporto.github.attention import GitHubAttentionReport
 from rapporto.github.backup import GitHubBackup
 from rapporto.github.model import GitHubInquiry
+from rapporto.util import to_mrkdwn
 
 organization_option = click.option("--organization", "--org", type=str, required=False)
 author_option = click.option("--author", type=str, required=False)
 when_option = click.option("--when", type=str, required=False)
+format_option = click.option("--format", "format_", type=str, required=False, default="markdown")
 repository_option = click.option("--repository", type=str, required=False)
 repositories_file_option = click.option("--repositories-file", type=Path, required=False)
 
@@ -30,23 +32,35 @@ def cli(ctx: click.Context):
 @organization_option
 @author_option
 @when_option
+@format_option
 def activity(
     organization: t.Optional[str] = None,
     author: t.Optional[str] = None,
     when: t.Optional[str] = None,
+    format_: t.Optional[str] = None,
 ):
     """
     Activities of individual authors.
     """
     inquiry = GitHubInquiry(organization=organization, author=author, created=when)
     report = GitHubActivityReport(inquiry=inquiry)
-    report.print()
+    print_output(report, format_)
+
+
+def print_output(report, format_):
+    if format_ == "markdown":
+        print(report.markdown)
+    elif format_ == "mrkdwn":
+        print(to_mrkdwn(report.markdown))
+    else:
+        raise NotImplementedError(f"Unknown output format: {format_}")
 
 
 @cli.command(aliases=["ci"])
 @repository_option
 @repositories_file_option
-def actions(repository: str, repositories_file: Path = None):
+@format_option
+def actions(repository: str, repositories_file: Path = None, format_: t.Optional[str] = None):
     """
     CI/GHA failures.
     """
@@ -60,19 +74,24 @@ def actions(repository: str, repositories_file: Path = None):
         )
         raise SystemExit(1) from ex
     report = GitHubActionsReport(inquiry=inquiry)
-    report.print()
+    print_output(report, format_)
 
 
 @cli.command(aliases=["att"])
 @organization_option
 @when_option
-def attention(organization: t.Optional[str] = None, when: t.Optional[str] = None):
+@format_option
+def attention(
+    organization: t.Optional[str] = None,
+    when: t.Optional[str] = None,
+    format_: t.Optional[str] = None,
+):
     """
     Important items that deserve attention.
     """
     inquiry = GitHubInquiry(organization=organization, created=when)
     report = GitHubAttentionReport(inquiry=inquiry)
-    report.print()
+    print_output(report, format_)
 
 
 @cli.command(